+++ app/controllers/users_controller.rb
@@ -1,4 +1,6 @@
 class UsersController < RestfulController
+  include RegexHelper
+
   # For rendering friends
   def index_scope(scope)
     requests = params[:requests]
@@ -8,26 +10,24 @@ class UsersController < RestfulController
       scope = current_user.incoming_friends
     end
 
-    return scope.order(last_name: :asc)
+    return scope
   end
 
   def create_friendship
-    #TODO Permit email and invite code params
+    #TODO: add `friend_finder` to permitted params
     authorize User, :create_friendship?
-    email = params[:email]
-    invite_code = params[:invite_code]
+    friend_finder = params[:friend_finder]
 
-    user = nil
-    user = User.find_by(email: email) if email.present?
-    user = User.find_by(invite_code: invite_code) if invite_code.present?
+    # `friend_finder` will be either the friend's email or friend code
+    user = User.find_by(email: friend_finder)
+    user = User.find_by(invite_code: friend_finder) if user.nil?
 
     if user.nil?
-      if invite_code.present? || !email.present?
+      if friend_finder.present? && friend_finder.match(VALID_EMAIL_REGEX).nil?
         render json: {error: { detail: "Unable to find a user by that invite code" }},
-               status: :forbidden
-      else
+               status: :not_found
+      elsif friend_finder.match(VALID_EMAIL_REGEX).present?
         # TODO: The user entered an email that isn't yet on the system, so let's send them an invite
-
       end
     else
       existing = Friendship.where("(user_id = #{current_user.id} AND friend_id = #{user.id}) OR
@@ -35,7 +35,7 @@ class UsersController < RestfulController
                      .first
 
       if existing.nil?
-        Friendship.create(user_id: current_user.id, friend_id: user.id)
+        Friendship.create!(user_id: current_user.id, friend_id: user.id)
         render nothing: true, status: :ok
       else
         # Friendship already exists
@@ -46,24 +46,25 @@ class UsersController < RestfulController
 
   def accept_friendship
     update_friendship do |friendship|
-      authorize friendship.friend, :accept_friendship?
-      friendship.update_attributes!(is_confirmed: true)
+      authorize friendship.user, :accept_friendship?
+      friendship.update_attributes!(is_confirmed: true) if friendship.present?
     end
   end
 
   def destroy_friendship
     update_friendship do |friendship|
-      authorize friendship.friend, :destroy_friendship?
-      friendship.destroy!
+      authorize User, :destroy_friendship?
+      friendship.destroy! if friendship.present?
     end
   end
 
   protected
 
   def update_friendship
-    friendship = Friendship.find(params[:friendship_id])
+    friendship = Friendship.find_by(id: params[:friendship_id])
+    current_user.friendship_to_authorize = friendship
+    yield friendship if block_given?
     if friendship.present?
-      yield friendship if block_given?
       render nothing: true, status: :no_content
     else
       render nothing: true, status: :not_found
